+++ interfaces.go
@@ -24,7 +24,7 @@ type Features interface {
 type Grabber interface {
 	Grab(ctx context.Context) (Sample, error)
 	GrabAll(ctx context.Context) ([]Sample, error)
-	SetImageKind(ImageKind)
+	SetComponent(Component)
 	Pause(ctx context.Context) error
 	Resume(ctx context.Context) error
 	Close() error